refactor: simplify interop imports
